add network selector
